Initial implementation
This is an initial implementation of ShareDB's [`MilestoneDB`][1]. The
base bones of `MongoMilestoneDB` are based on the [`sharedb-mongo`][2]
database adapter.

This adapter is a slight departure from other ShareDB code, as it is
written in ES7 for easier maintenance, especially when dealing with
asynchronous code, as this gives us access to the `async`/`await`
notation (although we can't realise its full potential due to needing
to deal with consumer callbacks to be consistent with the rest of
ShareDB).

The move to ES7 drops support for Node v6.

This library notably uses v2 of [`mongodb`][3] instead of the newer v3.
This is to stay consistent with `sharedb-mongo`, so in theory the same
config (or style of config) can be used in both `sharedb-mongo` and in
`sharedb-milestone-mongo`.

The version of [`istanbul`][4] is also bumped to the alpha version to
[support ES7 syntax][5]. We also move away from `jshint` and use
`eslint` with some more aggressive linting.

[1]: https://github.com/share/sharedb/pull/236
[2]: https://github.com/share/sharedb-mongo
[3]: https://mongodb.github.io/node-mongodb-native/
[4]: https://github.com/gotwarlost/istanbul
[5]: https://github.com/gotwarlost/istanbul/issues/733